requirements.txt
# FILE: api/requirements.txt

fastapi
uvicorn
python-multipart
pydantic
google-generativeai
PyMuPDF
python-dotenv